+++ e2e/BUILD.bazel
@@ -55,3 +55,18 @@ go_test(
         "@org_uber_go_zap//zaptest:go_default_library",
     ],
 )
+
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
