+++ e2e/testdata/TestCreatesInsecureCluster/creates_3-node_insecure_cluster.golden
@@ -27,9 +27,7 @@ spec:
   - args:
     - shell
     - -ecx
-    - '>- exec /cockroach/cockroach start --join=crdb-0.crdb.[sandbox_namespace]:26257,crdb-1.crdb.[sandbox_namespace]:26257,crdb-2.crdb.[sandbox_namespace]:26257
-      --advertise-host=$(hostname -f) --logtostderr=INFO --insecure --http-port=8080
-      --port=26257 --cache=25% --max-sql-memory=25%'
+    - '>- exec /cockroach/cockroach start --join=crdb-0.crdb.[sandbox_namespace]:26257,crdb-1.crdb.[sandbox_namespace]:26257,crdb-2.crdb.[sandbox_namespace]:26257 --advertise-host=$(hostname -f) --logtostderr=INFO --insecure --http-port=8080 --port=26257 --cache=25% --max-sql-memory=25%'
     env:
     - name: COCKROACH_CHANNEL
       value: kubernetes-operator
@@ -110,9 +108,7 @@ spec:
   - args:
     - shell
     - -ecx
-    - '>- exec /cockroach/cockroach start --join=crdb-0.crdb.[sandbox_namespace]:26257,crdb-1.crdb.[sandbox_namespace]:26257,crdb-2.crdb.[sandbox_namespace]:26257
-      --advertise-host=$(hostname -f) --logtostderr=INFO --insecure --http-port=8080
-      --port=26257 --cache=25% --max-sql-memory=25%'
+    - '>- exec /cockroach/cockroach start --join=crdb-0.crdb.[sandbox_namespace]:26257,crdb-1.crdb.[sandbox_namespace]:26257,crdb-2.crdb.[sandbox_namespace]:26257 --advertise-host=$(hostname -f) --logtostderr=INFO --insecure --http-port=8080 --port=26257 --cache=25% --max-sql-memory=25%'
     env:
     - name: COCKROACH_CHANNEL
       value: kubernetes-operator
@@ -193,9 +189,7 @@ spec:
   - args:
     - shell
     - -ecx
-    - '>- exec /cockroach/cockroach start --join=crdb-0.crdb.[sandbox_namespace]:26257,crdb-1.crdb.[sandbox_namespace]:26257,crdb-2.crdb.[sandbox_namespace]:26257
-      --advertise-host=$(hostname -f) --logtostderr=INFO --insecure --http-port=8080
-      --port=26257 --cache=25% --max-sql-memory=25%'
+    - '>- exec /cockroach/cockroach start --join=crdb-0.crdb.[sandbox_namespace]:26257,crdb-1.crdb.[sandbox_namespace]:26257,crdb-2.crdb.[sandbox_namespace]:26257 --advertise-host=$(hostname -f) --logtostderr=INFO --insecure --http-port=8080 --port=26257 --cache=25% --max-sql-memory=25%'
     env:
     - name: COCKROACH_CHANNEL
       value: kubernetes-operator
@@ -375,9 +369,7 @@ spec:
       - args:
         - shell
         - -ecx
-        - '>- exec /cockroach/cockroach start --join=crdb-0.crdb.[sandbox_namespace]:26257,crdb-1.crdb.[sandbox_namespace]:26257,crdb-2.crdb.[sandbox_namespace]:26257
-          --advertise-host=$(hostname -f) --logtostderr=INFO --insecure --http-port=8080
-          --port=26257 --cache=25% --max-sql-memory=25%'
+        - '>- exec /cockroach/cockroach start --join=crdb-0.crdb.[sandbox_namespace]:26257,crdb-1.crdb.[sandbox_namespace]:26257,crdb-2.crdb.[sandbox_namespace]:26257 --advertise-host=$(hostname -f) --logtostderr=INFO --insecure --http-port=8080 --port=26257 --cache=25% --max-sql-memory=25%'
         env:
         - name: COCKROACH_CHANNEL
           value: kubernetes-operator
