+++ hack/boilerplate/BUILD.bazel
@@ -1,0 +1,7 @@
+package(default_visibility = ["//visibility:public"])
+
+filegroup(
+    name = "boilerplate-srcs",
+    srcs = glob(["**"]),
+)
+
