+++ hack/requirements.txt
@@ -1,0 +1,8 @@
+attrs==19.3.0
+more-itertools==8.2.0
+packaging==20.3
+pluggy==0.13.1
+psycopg2-binary==2.8.5
+py==1.8.1
+pytest==5.4.1
+wcwidth==0.1.9
